+++ src/form/b-icon-button/b-icon-button.ts
@@ -1,0 +1,13 @@
+/*!
+ * V4Fire Client Core
+ * https://github.com/V4Fire/Client
+ *
+ * Released under the MIT license
+ * https://github.com/V4Fire/Client/blob/master/LICENSE
+ */
+
+import bButton, { component } from 'form/b-button/b-button';
+export * from 'form/b-button/b-button';
+
+@component({flyweight: true})
+export default class bIconButton<T extends object = Dictionary> extends bButton<T> {}